gateway/agent-service/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
httpx>=0.28.0,<1.0.0
pyyaml>=6.0.2,<7.0.0
pydantic>=2.10.0,<3.0.0
